requirements.txt
langchain
streamlit
tiktoken
pinecone-client
streamlit_chat